+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 ### Added
 ### Fixed
+
+- Fixed issue where `List` holding `Schema` type items would lose items to validation warnings (#18)
+
 ### Changed
 ### Deprecated
 ### Removed